+++ src/resources/views/create-post.blade.php
@@ -19,9 +19,33 @@
                 <option value="0" @if (old('published') == '0') {{ 'selected' }} @endif>Draft</option>
             </select>
 
-            <label for="banner_image_url">Banner Image (eg: 1234567890.png) <i class="fas fa-chevron-right"></i> <a
-                    href="{{ route('images') }}">Image Library</a></label>
-            <input type="text" class="form-control" name="banner_image_url" value="{{ old('banner_image_url') }}">
+            <label for="banner_image_url">Banner Image</label>
+            <input type="text" class="form-control" name="banner_image_url" id="banner_image_url"
+                value="{{ old('banner_image_url') }}" hidden>
+
+            <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#imageGallery">
+                Select Banner Image
+            </button>
+
+            <img src="{{ asset('pics/banner.jpg') }}" id="banner_image_preview">
+
+            <div class="modal fade" id="imageGallery" tabindex="-1" aria-labelledby="imageGallery" aria-hidden="true">
+                <div class="modal-dialog modal-dialog-scrollable">
+                    <div class="modal-content">
+                        <div class="modal-header">
+                            <h5 class="modal-title" id="imageGallery">Select Banner Image</h5>
+                            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                        </div>
+                        <div class="modal-body">
+                            <p>Go to the <a href="{{ route('images') }}">Image Library</a> to manage images.</p>
+                            @include('partials.image-gallery')
+                        </div>
+                        <div class="modal-footer">
+                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
 
             <label for="category_id">Category</label>
             <select class="form-control" name="category_id">