+++ src/resources/views/edit-post.blade.php
@@ -26,16 +26,34 @@
                 </option>
             </select>
 
-            <label for="banner_image_url">Banner Image (eg: 1234567890.png) <i class="fas fa-chevron-right"></i> <a
-                    href="{{ route('images') }}">Image Library</a></label>
-            <input type="text" class="form-control" name="banner_image_url"
-                value="{{ old('banner_image_url', $post->banner_image_url) }}">
-            @if (!file_exists(public_path("storage/images/posts/$post->banner_image_url")) || $post->banner_image_url == null)
-                <small class="text-danger">Warning: The image provided could not be found. The default placeholder image
-                    will
-                    be
-                    used.</small>
-            @endif
+            <label for="banner_image_url">Banner Image</label>
+            <input type="text" class="form-control" name="banner_image_url" id="banner_image_url"
+                value="{{ old('banner_image_url', $post->banner_image_url) }}" hidden>
+
+            <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#imageGallery">
+                Select Banner Image
+            </button>
+
+            <img src="{{ \App\Http\Controllers\PostController::getImageAssetPath($post->banner_image_url) }}"
+                id="banner_image_preview">
+
+            <div class="modal fade" id="imageGallery" tabindex="-1" aria-labelledby="imageGallery" aria-hidden="true">
+                <div class="modal-dialog modal-dialog-scrollable">
+                    <div class="modal-content">
+                        <div class="modal-header">
+                            <h5 class="modal-title" id="imageGallery">Select Banner Image</h5>
+                            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                        </div>
+                        <div class="modal-body">
+                            <p>Go to the <a href="{{ route('images') }}">Image Library</a> to manage images.</p>
+                            @include('partials.image-gallery')
+                        </div>
+                        <div class="modal-footer">
+                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
 
             <label for="category_id">Category</label>
             <select class="form-select" name="category_id">
@@ -51,10 +69,13 @@
 
             <label for="post">Post</label>
             <textarea name="post" class="form-control" id="post" rows="6">{{ old('post', $post->post) }}</textarea>
-            <a href="https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet" target="_blank">Markdown Cheat
-                Sheet</a>
 
-            <br />
+            <p>
+                <a href="https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet" target="_blank">
+                    Markdown Cheat Sheet
+                </a>
+            </p>
+
             <input type="submit" class="btn btn-primary" value="Update Post">
         </form>
     </div>