+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Next Release
 
+- Adds an image gallery selection to the create and edit post screens. Previously, you'd need to visit the Image Library, copy the image name, go back to your post, paste in the image name to a field, and be on your way
 - Fixes a bug where retrieving posts by category didn't work due to a variable name error
-- Fixes various other bugs and makes additional improvements
+- Various other improvements and bug fixes
 
 ## v0.12.1 (2022-09-17)